[fix] 함수 분리 및 코드 정리
