+++ web_frontend/src/constants/CommentTexts.tsx
@@ -25,12 +25,15 @@ export const calText2 = [
 ];
 
 export const nutText = {
-  0: `단백질이 너무 많습니다. 과도한 단백질 섭취는 신장 부담을 초래하고, 뼈 밀도 감소, 심장 질환 등의 건강 문제를 유발할 수 있습니다.`,
+  0: `단백질이 너무 많습니다.\n`,
   1: `섭취한 단백질이 부족합니다. 육류, 계랸, 콩류를 드셔보는 것도 좋을 것 같습니다.`,
-  2: `과다한 탄수화물 섭취는 체중 증가, 인슐린 저항성, 당뇨병 등의 건강 문제를 유발할 수 있습니다.`,
+  2: `탄수화물 섭취가 많습니다.\n`,
   3: `탄수화물 섭취가 부족합니다. 탄수화물이 부족한 경우 에너지 부족, 불규칙한 식습관, 신경 쇠약 등의 증상이 나타날 수 있습니다.`,
-  4: `과다한 지방 섭취는 고지혈증, 동맥경화, 심혈관 질환 등의 건강 문제를 유발할 수 있습니다. 따라서, 권장량을 초과하지 않도록 주의해야 합니다`,
-  5: `지방 섭취가 부족합니다. 지방이 부족한 경우 영양 실조, 비만 등의 증상이 나타날 수 있습니다.`,
+  4: `과다한 지방 섭취는 고지혈증, 동맥경화, 심혈관 질환 등의 건강 문제를 유발할 수 있습니다.`,
+  5: `지방 섭취가 부족합니다. \n`,
   6: `콜레스테롤이 너무 많습니다. 콜레스테롤이 많으면 동맥경화, 심장 질환, 뇌졸중 등의 위험이 증가할 수 있습니다.`,
   7: `나트륨이 많습니다. 나트륨이 많을시에 혈압, 심장 질환, 뇌졸중 등의 위험이 증가할 수 있습니다. `,
+  8: `단백질은 적당합니다 \n`,
+  9: `탄수화물은 적당합니다 \n`,
+  10: `지방이 적당합니다 \n`,
 };
